Bump versions of all packages that define "c" methods
The Windows and Mac binary packages currently available in BioC 3.4 were all
built with R 3.3.2. Unfortunately, because of a change in the signature of
the c() generic introduced in R 3.3.3, binary packages built with R 3.3.2 and
defining "c" methods are incompatible with R 3.3.3 and need to be rebuilt with
R 3.3.3. Now that our Windows and Mac build machines for BioC 3.4 have been
updated to use R 3.3.3, these version bumps will trigger the rebuilt of the
broken binaries and the propagation of the new binaries to the public repos.
IMPORTANT NOTE: R 3.3.2 users will get a spurious warning that these new
binaries have been built with R 3.3.3 when loading them. It seems safe to
ignore it though.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_3_4/madman/Rpacks/GGtools@127452 bc3139a8-67e5-0310-9ffc-ced21a209358